Add basic notebook layout and move actions into `positronCustomViews.ts` instead of polluting `layoutActions.ts`
